+++ exam/data/db/schema.sql
@@ -5,9 +5,9 @@
 CREATE SEQUENCE authentication_id;
 
 CREATE TABLE authentication (
-  id            numeric(10, 0) NOT NULL DEFAULT nextval('authentication_id'),
+  id            bigint NOT NULL DEFAULT nextval('authentication_id'),
   token         varchar(255) NOT NULL,
-  user_account  numeric(10, 0) NOT NULL,
+  user_account  bigint NOT NULL,
 
   CONSTRAINT id_pk PRIMARY KEY (id),
   CONSTRAINT token_uq UNIQUE (token)
